requirements.txt
flask==3.1.0
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0
requests==2.32.3
pandas==2.2.3